Cambio de distribucion de ficheros y hechos algunos ejemplos mas
