+++ pruebas/pruebas_calc_grafica.py
